Move lyric caret into single layer.
