+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Rows/EditLyricRow.cs
@@ -10,15 +10,12 @@ using osu.Framework.Graphics.Containers;
 using osu.Framework.Graphics.Cursor;
 using osu.Framework.Graphics.Shapes;
 using osu.Framework.Graphics.UserInterface;
-using osu.Framework.Input.Events;
 using osu.Game.Graphics;
 using osu.Game.Graphics.Sprites;
 using osu.Game.Graphics.UserInterface;
 using osu.Game.Overlays;
 using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Lyrics;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.CaretPosition;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.Carets;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.FixedInfo;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.Parts;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows.Components.SubInfo;
@@ -284,15 +281,11 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows
             private readonly EditorLyricPiece lyricPiece;
 
             private readonly Container timeTagContainer;
-            private readonly Container<DrawableCaret> caretContainer;
 
             private readonly IBindable<LyricEditorMode> bindableMode = new Bindable<LyricEditorMode>();
 
             [Resolved]
             private ILyricsChangeHandler lyricsChangeHandler { get; set; }
-
-            [Resolved]
-            private ILyricCaretState lyricCaretState { get; set; }
 
             public Lyric Lyric { get; }
 
@@ -309,7 +302,7 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows
                     {
                         RelativeSizeAxes = Axes.Both,
                     },
-                    caretContainer = new Container<DrawableCaret>
+                    new CaretLayer(lyric)
                     {
                         RelativeSizeAxes = Axes.Both,
                     }
@@ -322,111 +315,9 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows
 
                 bindableMode.BindValueChanged(e =>
                 {
-                    // initial default caret.
-                    InitializeCaret(e.NewValue);
-
                     // Initial blueprint container.
                     InitializeBlueprint(e.NewValue);
                 });
-            }
-
-            protected override bool OnMouseMove(MouseMoveEvent e)
-            {
-                if (!lyricCaretState.CaretEnabled)
-                    return false;
-
-                var mode = bindableMode.Value;
-                float position = ToLocalSpace(e.ScreenSpaceMousePosition).X;
-
-                switch (mode)
-                {
-                    case LyricEditorMode.View:
-                        break;
-
-                    case LyricEditorMode.Manage:
-                        int cuttingLyricStringIndex = Math.Clamp(TextIndexUtils.ToStringIndex(lyricPiece.GetHoverIndex(position)), 0, Lyric.Text.Length - 1);
-                        lyricCaretState.MoveHoverCaretToTargetPosition(new TextCaretPosition(Lyric, cuttingLyricStringIndex));
-                        break;
-
-                    case LyricEditorMode.Typing:
-                        int typingStringIndex = TextIndexUtils.ToStringIndex(lyricPiece.GetHoverIndex(position));
-                        lyricCaretState.MoveHoverCaretToTargetPosition(new TextCaretPosition(Lyric, typingStringIndex));
-                        break;
-
-                    case LyricEditorMode.Language:
-                        break;
-
-                    case LyricEditorMode.EditRuby:
-                        lyricCaretState.MoveHoverCaretToTargetPosition(new NavigateCaretPosition(Lyric));
-                        break;
-
-                    case LyricEditorMode.EditRomaji:
-                        lyricCaretState.MoveHoverCaretToTargetPosition(new NavigateCaretPosition(Lyric));
-                        break;
-
-                    case LyricEditorMode.CreateTimeTag:
-                        var textIndex = lyricPiece.GetHoverIndex(position);
-                        lyricCaretState.MoveHoverCaretToTargetPosition(new TimeTagIndexCaretPosition(Lyric, textIndex));
-                        break;
-
-                    case LyricEditorMode.RecordTimeTag:
-                        var timeTag = lyricPiece.GetHoverTimeTag(position);
-                        lyricCaretState.MoveHoverCaretToTargetPosition(new TimeTagCaretPosition(Lyric, timeTag));
-                        break;
-
-                    case LyricEditorMode.AdjustTimeTag:
-                    case LyricEditorMode.EditNote:
-                    case LyricEditorMode.Singer:
-                        lyricCaretState.MoveHoverCaretToTargetPosition(new NavigateCaretPosition(Lyric));
-                        break;
-
-                    default:
-                        throw new ArgumentOutOfRangeException(nameof(mode));
-                }
-
-                return base.OnMouseMove(e);
-            }
-
-            protected override void OnHoverLost(HoverLostEvent e)
-            {
-                if (!lyricCaretState.CaretEnabled)
-                    return;
-
-                // lost hover caret and time-tag caret
-                lyricCaretState.ClearHoverCaretPosition();
-                base.OnHoverLost(e);
-            }
-
-            protected override bool OnClick(ClickEvent e)
-            {
-                if (!lyricCaretState.CaretEnabled)
-                    return false;
-
-                // place hover caret to target position.
-                var position = lyricCaretState.BindableHoverCaretPosition.Value;
-                if (position == null)
-                    return false;
-
-                lyricCaretState.MoveCaretToTargetPosition(position);
-
-                return true;
-            }
-
-            protected override bool OnDoubleClick(DoubleClickEvent e)
-            {
-                var mode = bindableMode.Value;
-                var position = lyricCaretState.BindableCaretPosition.Value;
-
-                switch (position)
-                {
-                    case TextCaretPosition textCaretPosition:
-                        if (mode == LyricEditorMode.Manage)
-                            lyricsChangeHandler.Split(textCaretPosition.Index);
-                        return true;
-
-                    default:
-                        return false;
-                }
             }
 
             [BackgroundDependencyLoader]
@@ -456,60 +347,6 @@ namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Rows
                         LyricEditorMode.AdjustTimeTag => new TimeTagBlueprintContainer(lyric),
                         _ => null
                     };
-            }
-
-            protected void InitializeCaret(LyricEditorMode mode)
-            {
-                caretContainer.Clear();
-
-                // create preview and real caret
-                addCaret(false);
-                addCaret(true);
-
-                void addCaret(bool isPreview)
-                {
-                    var caret = createCaret(mode, isPreview);
-                    if (caret == null)
-                        return;
-
-                    caret.Hide();
-
-                    caretContainer.Add(caret);
-                }
-
-                static DrawableCaret createCaret(LyricEditorMode mode, bool isPreview)
-                {
-                    switch (mode)
-                    {
-                        case LyricEditorMode.View:
-                            return null;
-
-                        case LyricEditorMode.Manage:
-                            return new DrawableLyricSplitterCaret(isPreview);
-
-                        case LyricEditorMode.Typing:
-                            return new DrawableLyricInputCaret(isPreview);
-
-                        case LyricEditorMode.Language:
-                        case LyricEditorMode.EditRuby:
-                        case LyricEditorMode.EditRomaji:
-                            return null;
-
-                        case LyricEditorMode.CreateTimeTag:
-                            return new DrawableTimeTagEditCaret(isPreview);
-
-                        case LyricEditorMode.RecordTimeTag:
-                            return new DrawableTimeTagRecordCaret(isPreview);
-
-                        case LyricEditorMode.AdjustTimeTag:
-                        case LyricEditorMode.EditNote:
-                        case LyricEditorMode.Singer:
-                            return null;
-
-                        default:
-                            throw new ArgumentOutOfRangeException(nameof(mode));
-                    }
-                }
             }
 
             protected void UpdateTimeTags()
